requirements.txt
streamlit
pytesseract
opencv-python-headless
numpy
Pillow
Flask 